Refactor: refactor directory structure
