+++ src/main/java/org/prgrms/kdt/etc/OrderTester.java
@@ -3,8 +3,8 @@ package org.prgrms.kdt.etc;
 import org.prgrms.kdt.configuration.AppConfiguration;
 import org.prgrms.kdt.order.OrderItem;
 import org.prgrms.kdt.order.OrderService;
-import org.prgrms.kdt.voucher.FixedAmountVoucher;
-import org.prgrms.kdt.voucher.VoucherRepository;
+import org.prgrms.kdt.voucher.model.FixedAmountVoucher;
+import org.prgrms.kdt.voucher.repository.VoucherRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.ansi.AnsiOutput;
