+++ src/main/java/org/prgrms/kdt/order/OrderService.java
@@ -1,6 +1,6 @@
 package org.prgrms.kdt.order;
 
-import org.prgrms.kdt.voucher.VoucherService;
+import org.prgrms.kdt.voucher.service.VoucherService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
