--- src/main/java/org/prgrms/kdt/voucher/VoucherList.java
@@ -1,11 +1,0 @@
-package org.prgrms.kdt.voucher;
-
-import java.util.List;
-
-public class VoucherList {
-    private List<Voucher> voucherList;
-
-    public List<Voucher> getVoucherList() {
-        return voucherList;
-    }
-}
