+++ src/main/java/org/prgrms/kdt/voucher/VoucherType.java
@@ -1,15 +1,15 @@
 package org.prgrms.kdt.voucher;
 
 public enum VoucherType {
-    FixedAmountVoucher(1),
-    PercentDiscountVoucher(2);
+    FixedAmountVoucher,
+    PercentDiscountVoucher
 
-    private final int voucherType;
-    VoucherType(int voucherType) {
-        this.voucherType = voucherType;
-    }
-
-    public int typeNum() {
-        return voucherType;
-    }
+//    private final int voucherType;
+//    VoucherType(int voucherType) {
+//        this.voucherType = voucherType;
+//    }
+//
+//    public int typeNum() {
+//        return voucherType;
+//    }
 }
