+++ src/main/java/org/prgrms/kdt/voucher/model/PercentDiscountVoucher.java
@@ -1,6 +1,7 @@
-package org.prgrms.kdt.voucher;
+package org.prgrms.kdt.voucher.model;
 
 import lombok.ToString;
+import org.prgrms.kdt.voucher.VoucherType;
 
 import java.util.UUID;
