+++ src/test/java/org/prgrms/kdt/KdtSpringContextTests.java
@@ -5,10 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.prgrms.kdt.order.OrderItem;
 import org.prgrms.kdt.order.OrderService;
 import org.prgrms.kdt.order.OrderStatus;
-import org.prgrms.kdt.voucher.FixedAmountVoucher;
-import org.prgrms.kdt.voucher.MemoryVoucherRepository;
-import org.prgrms.kdt.voucher.VoucherRepository;
-import org.prgrms.kdt.voucher.VoucherService;
+import org.prgrms.kdt.voucher.model.FixedAmountVoucher;
+import org.prgrms.kdt.voucher.repository.VoucherRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
