+++ src/test/java/org/prgrms/kdt/voucher/JdbcVoucherRepositoryTest.java
@@ -3,6 +3,9 @@ package org.prgrms.kdt.voucher;
 import com.zaxxer.hikari.HikariDataSource;
 import org.junit.jupiter.api.*;
 import org.prgrms.kdt.customer.repository.CustomerNamedJdbcRepository;
+import org.prgrms.kdt.voucher.model.FixedAmountVoucher;
+import org.prgrms.kdt.voucher.model.Voucher;
+import org.prgrms.kdt.voucher.repository.JdbcVoucherRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
